Integrate find-novel-kmers
